mv cylinder3d_module to backbone
