+++ configs/cylinder3d/cylinder3d_4xb4_3x_semantickitti.py
@@ -28,6 +28,10 @@ param_scheduler = [
         gamma=0.1)
 ]
 
+train_dataloader = dict(
+    batch_size=4,
+)
+
 # Default setting for scaling LR automatically
 #   - `enable` means enable scaling LR automatically
 #       or not by default.
